requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-annoying==0.10.4
django-appconf==1.0.2
django-compressor==2.2
django-material==1.2.2
django-phonenumber-field==2.0.0
django-sass-processor==0.5.8
idna==2.6
libsass==0.13.7
phonenumberslite==8.9.0
pushbullet.py==0.11.0
python-magic==0.4.15
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
